requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
gunicorn==21.2.0

# Data & ML
numpy>=1.26
tensorflow>=2.16
scikit-learn>=1.3.3
pandas==2.2.3



# Stock Data
yfinance==0.2.32
ta==0.11.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlalchemy==2.0.23

# Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3


